Add dependencies eslint prop-types react-router react-router-dom
